Initialize 処理を基底クラスに分ける
