+++ Assets/ArtNet/Editor/UnityRecorder/Input/TargetedPropertyDrawer.cs
@@ -1,0 +1,32 @@
+using System.Reflection;
+using UnityEditor;
+using UnityEngine;
+
+namespace ArtNet.UnityRecorder.Input
+{
+    public class TargetedPropertyDrawer<T> : PropertyDrawer where T : class
+    {
+        protected T Target;
+
+        protected virtual void Initialize(SerializedProperty prop)
+        {
+            var path = prop.propertyPath.Split('.');
+            object obj = prop.serializedObject.targetObject;
+
+            foreach (var pathNode in path)
+                obj = GetSerializedField(obj, pathNode).GetValue(obj);
+
+            Target = obj as T;
+        }
+
+        private static FieldInfo GetSerializedField(object target, string pathNode)
+        {
+            return target.GetType().GetField(pathNode, BindingFlags.Instance | BindingFlags.NonPublic | BindingFlags.Public);
+        }
+
+        public override float GetPropertyHeight(SerializedProperty property, GUIContent label)
+        {
+            return 0.0f;
+        }
+    }
+}
